almost fully moved out of main
